Separate dropdown handlers into different files
